Add environment variable loader to `watch` task
This allows local environment variables to be defined in a `.env` file in the project to be used for local development. This is easier to manage per-project than setting global env vars (which apply to all projects) or defining vars at runtime using the command line (which is cumbersome to repeat).

Additionally, allowing the file to be specified allows for the easy switching between multiple configuration setups.